イベント一覧系への遷移時にデータ取得完了して描画してから遷移するようにした
