+++ src/EventList.vue
@@ -10,12 +10,14 @@
         過去の開催イベントは<router-link to="/past_events" class="text-link">こちら</router-link>
       </p>
     </section>
-    <event-list-view></event-list-view>
+    <event-list-view :event-items="EventItems"></event-list-view>
     <p class="coming-soon">coming soon...</p>
     <top-footer></top-footer>
   </div>
 </template>
 <script>
+import firebase from 'firebase';
+import firestore from './assets/javascript/firebase';
 import ContentTitle2 from './components/ContentTitle2';
 import EventListView from './components/EventListView';
 import TopFooter from './components/layouts/TopFooter';
@@ -26,6 +28,62 @@ export default {
     EventListView,
     ContentTitle2,
     TopFooter,
+  },
+  data() {
+    return {
+      EventItems: [],
+    };
+  },
+  beforeRouteEnter(route, redirect, next) {
+    firestore.collection('events').where('begin_datetime', '>=', firebase.firestore.Timestamp.now()).orderBy('begin_datetime').get()
+      .then((querySnapshot) => {
+        const events = [];
+        querySnapshot.forEach((doc) => {
+          events.push(doc);
+        });
+        next((vm) => {
+          vm.dataSet(events);
+        });
+      });
+  },
+  methods: {
+    fromTimeStampToDate(date) {
+      const WeekChars = ['日', '月', '火', '水', '木', '金', '土'];
+      const d = new Date(date.seconds * 1000);
+      const year = d.getFullYear();
+      const month = (`0${d.getMonth() + 1}`).slice(-2);
+      const day = (`0${d.getDate()}`).slice(-2);
+      const wday = d.getDay();
+      return `${year}.${month}.${day}( ${WeekChars[wday]} )`;
+    },
+    fromTimeStampToTime(date) {
+      const d = new Date(date.seconds * 1000);
+      const hour = (`0${d.getHours()}`).slice(-2);
+      const min = (`0${d.getMinutes()}`).slice(-2);
+      return `${hour}:${min}`;
+    },
+    getNowFormattedFirebase() {
+      return firebase.firestore.Timestamp.now();
+    },
+    replaceNewlineTag(str) {
+      return str.replace(/<br>/g, ' ');
+    },
+    dataSet(data) {
+      const eventListItems = [];
+      data.forEach((doc) => {
+        eventListItems.push(
+          {
+            id: doc.id,
+            title: doc.data().title,
+            date: this.fromTimeStampToDate(doc.data().begin_datetime),
+            time: this.fromTimeStampToTime(doc.data().begin_datetime),
+            image_path: `/static/img/${doc.data().image_path}`,
+            description: this.replaceNewlineTag(doc.data().description),
+            link_url: doc.data().link_url,
+          });
+      });
+      this.EventItems = eventListItems;
+    },
   },
 };
 </script>
